refactor: 테스트 격리, 공통 fixture 를 갖는 base test 정의
- Base test class 인 AcceptanceTest 정의
- 기존 인수 테스트가 AcceptanceTest 를 상속하도록 변경
